// src/api/openapi/auto-generated/models/APIEIModule.ts
/* tslint:disable */
/* eslint-disable */ // @ts-nocheck
/**
 * Jupiter API
 * Welcome to Jupiter\'s ClimateScore Global API
 *
 * The version of the OpenAPI document: 2.2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * An enumeration.
 * @export
 */
export const APIEIModule = {
    acuteCombinedFlood: 'acute_combined_flood',
    acuteFire: 'acute_fire',
    acuteWind: 'acute_wind',
    utilitiesCooling: 'utilities_cooling',
    utilitiesWater: 'utilities_water',
    productivityHeat: 'productivity_heat',
    financial: 'financial',
    unknownDefaultOpenApi: '11184809'
} as const;
export type APIEIModule = typeof APIEIModule[keyof typeof APIEIModule];


export function instanceOfAPIEIModule(value: any): boolean {
    for (const key in APIEIModule) {
        if (Object.prototype.hasOwnProperty.call(APIEIModule, key)) {
            if ((APIEIModule as Record<string, APIEIModule>)[key] === value) {
                return true;
            }
        }
    }
    return false;
}

export function APIEIModuleFromJSON(json: any): APIEIModule {
    return APIEIModuleFromJSONTyped(json, false);
}

export function APIEIModuleFromJSONTyped(json: any, ignoreDiscriminator: boolean): APIEIModule {
    return json as APIEIModule;
}

export function APIEIModuleToJSON(value?: APIEIModule | null): any {
    return value as any;
}