// src/api/openapi/auto-generated/models/PerilsResultSetOptions.ts
export interface PerilsResultSetOptions {
    /**
     * 
     * @type {string}
     * @memberof PerilsResultSetOptions
     */
    dataVersion?: PerilsResultSetOptionsDataVersionEnum;
    /**
     * 
     * @type {Array<string>}
     * @memberof PerilsResultSetOptions
     */
    scenarios?: Array<PerilsResultSetOptionsScenariosEnum>;
    /**
     * 
     * @type {Array<number>}
     * @memberof PerilsResultSetOptions
     */
    years?: Array<PerilsResultSetOptionsYearsEnum>;
    /**
     * 
     * @type {Array<string>}
     * @memberof PerilsResultSetOptions
     */
    perils?: Array<PerilsResultSetOptionsPerilsEnum>;
    /**
     * 
     * @type {FloodDefenseOptions}
     * @memberof PerilsResultSetOptions
     */
    floodDefense?: FloodDefenseOptions;
    /**
     * 
     * @type {string}
     * @memberof PerilsResultSetOptions
     */
    type: PerilsResultSetOptionsTypeEnum;
}


/**
 * @export
 */
export const PerilsResultSetOptionsDataVersionEnum = {
    _26: '2.6',
    _262: '2.6.2',
    _3: '3',
    _30: '3.0',
    _300: '3.0.0',
    _31: '3.1',
    _310: '3.1.0',
    _32: '3.2',
    _320: '3.2.0',
    _330: '3.3.0',
    unknownDefaultOpenApi: '11184809'
} as const;
export type PerilsResultSetOptionsDataVersionEnum = typeof PerilsResultSetOptionsDataVersionEnum[keyof typeof PerilsResultSetOptionsDataVersionEnum];

/**
 * @export
 */
export const PerilsResultSetOptionsScenariosEnum = {
    baseline: 'baseline',
    ssp126: 'ssp126',
    ssp245: 'ssp245',
    ssp585: 'ssp585',
    unknownDefaultOpenApi: '11184809'
} as const;
export type PerilsResultSetOptionsScenariosEnum = typeof PerilsResultSetOptionsScenariosEnum[keyof typeof PerilsResultSetOptionsScenariosEnum];

/**
 * @export
 */
export const PerilsResultSetOptionsYearsEnum = {
    NUMBER_1995: 1995,
    NUMBER_2020: 2020,
    NUMBER_2025: 2025,
    NUMBER_2030: 2030,
    NUMBER_2035: 2035,
    NUMBER_2040: 2040,
    NUMBER_2045: 2045,
    NUMBER_2050: 2050,
    NUMBER_2055: 2055,
    NUMBER_2060: 2060,
    NUMBER_2065: 2065,
    NUMBER_2070: 2070,
    NUMBER_2075: 2075,
    NUMBER_2080: 2080,
    NUMBER_2085: 2085,
    NUMBER_2090: 2090,
    NUMBER_2095: 2095,
    NUMBER_2100: 2100,
    NUMBER_unknown_default_open_api: 11184809
} as const;
export type PerilsResultSetOptionsYearsEnum = typeof PerilsResultSetOptionsYearsEnum[keyof typeof PerilsResultSetOptionsYearsEnum];

/**
 * @export
 */
export const PerilsResultSetOptionsPerilsEnum = {
    cold: 'cold',
    combinedFlood: 'combinedFlood',
    drought: 'drought',
    fire: 'fire',
    hail: 'hail',
    heat: 'heat',
    precip: 'precip',
    wind: 'wind',
    subsidence: 'subsidence',
    coastalFlood: 'coastalFlood',
    fluvialFlood: 'fluvialFlood',
    pluvialFlood: 'pluvialFlood',
    subsidence: 'subsidence',
    unknownDefaultOpenApi: '11184809'
} as const;
export type PerilsResultSetOptionsPerilsEnum = typeof PerilsResultSetOptionsPerilsEnum[keyof typeof PerilsResultSetOptionsPerilsEnum];

/**
 * @export
 */
export const PerilsResultSetOptionsTypeEnum = {
    perils: 'perils',
    unknownDefaultOpenApi: '11184809'
} as const;
export type PerilsResultSetOptionsTypeEnum = typeof PerilsResultSetOptionsTypeEnum[keyof typeof PerilsResultSetOptionsTypeEnum];