// src/api/openapi/auto-generated/models/PipelineStatus.ts
/* tslint:disable */
/* eslint-disable */ // @ts-nocheck
/**
 * Jupiter API
 * Welcome to Jupiter\'s ClimateScore Global API
 *
 * The version of the OpenAPI document: 2.2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * An enumeration.
 * @export
 */
export const PipelineStatus = {
    pending: 'pending',
    completed: 'completed',
    failed: 'failed',
    unknownDefaultOpenApi: '11184809'
} as const;
export type PipelineStatus = typeof PipelineStatus[keyof typeof PipelineStatus];


export function instanceOfPipelineStatus(value: any): boolean {
    for (const key in PipelineStatus) {
        if (Object.prototype.hasOwnProperty.call(PipelineStatus, key)) {
            if ((PipelineStatus as Record<string, PipelineStatus>)[key] === value) {
                return true;
            }
        }
    }
    return false;
}

export function PipelineStatusFromJSON(json: any): PipelineStatus {
    return PipelineStatusFromJSONTyped(json, false);
}

export function PipelineStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): PipelineStatus {
    return json as PipelineStatus;
}

export function PipelineStatusToJSON(value?: PipelineStatus | null): any {
    return value as any;
}