// src/api/openapi/auto-generated/models/ScoresResultSetOptions.ts
export interface ScoresResultSetOptions {
    /**
     * 
     * @type {string}
     * @memberof ScoresResultSetOptions
     */
    dataVersion?: ScoresResultSetOptionsDataVersionEnum;
    /**
     * 
     * @type {boolean}
     * @memberof ScoresResultSetOptions
     */
    includeBenchmarks?: boolean;
    /**
     * 
     * @type {Array<JupiterApiEnumsBenchmarkingLevel>}
     * @memberof ScoresResultSetOptions
     */
    benchmarkLevels?: Array<JupiterApiEnumsBenchmarkingLevel>;
    /**
     * 
     * @type {Array<string>}
     * @memberof ScoresResultSetOptions
     */
    perils?: Array<ScoresResultSetOptionsPerilsEnum>;
    /**
     * 
     * @type {string}
     * @memberof ScoresResultSetOptions
     */
    type: ScoresResultSetOptionsTypeEnum;
}


/**
 * @export
 */
export const ScoresResultSetOptionsDataVersionEnum = {
    _26: '2.6',
    _262: '2.6.2',
    _3: '3',
    _30: '3.0',
    _300: '3.0.0',
    _31: '3.1',
    _310: '3.1.0',
    _32: '3.2',
    _320: '3.2.0',
    _330: '3.3.0',
    unknownDefaultOpenApi: '11184809'
} as const;
export type ScoresResultSetOptionsDataVersionEnum = typeof ScoresResultSetOptionsDataVersionEnum[keyof typeof ScoresResultSetOptionsDataVersionEnum];

/**
 * @export
 */
export const ScoresResultSetOptionsPerilsEnum = {
    cold: 'cold',
    combinedFlood: 'combinedFlood',
    drought: 'drought',
    fire: 'fire',
    hail: 'hail',
    heat: 'heat',
    precip: 'precip',
    wind: 'wind',
    subsidence: 'subsidence',
    allPerils: 'allPerils',
    unknownDefaultOpenApi: '11184809'
} as const;
export type ScoresResultSetOptionsPerilsEnum = typeof ScoresResultSetOptionsPerilsEnum[keyof typeof ScoresResultSetOptionsPerilsEnum];

/**
 * @export
 */
export const ScoresResultSetOptionsTypeEnum = {
    scores: 'scores',
    unknownDefaultOpenApi: '11184809'
} as const;
export type ScoresResultSetOptionsTypeEnum = typeof ScoresResultSetOptionsTypeEnum[keyof typeof ScoresResultSetOptionsTypeEnum];